Refactor and fix code quality
