+++ src/qililab/qprogram/experiment.py
@@ -21,14 +21,31 @@ from qililab.yaml import yaml
 
 @yaml.register_class
 class Experiment(StructuredProgram):
-    def __init__(self) -> None:
-        super().__init__()
+    """Represents an experiment.
+
+    This class allows setting platform parameters and executing quantum programs.
+    """
 
     def set_parameter(self, alias: str, parameter: Parameter, value: int | float | int):
-        """ """
+        """Set a platform parameter.
+
+        Appends a SetParameter operation to the active block of the experiment.
+
+        Args:
+            alias (str): The alias for the platform component.
+            parameter (Parameter): The parameter to set.
+            value (int | float): The value to set for the parameter.
+        """
         operation = SetParameter(alias=alias, parameter=parameter, value=value)
         self._active_block.append(operation)
 
     def execute_qprogram(self, qprogram: QProgram):
+        """Execute a quantum program within the experiment.
+
+        Appends an ExecuteQProgram operation to the active block of the experiment.
+
+        Args:
+            qprogram (QProgram): The quantum program to be executed.
+        """
         operation = ExecuteQProgram(qprogram=qprogram)
         self._active_block.append(operation)
